Removed redundant telemetry parser
Removed redundant telm parser and fixed typos while adding debug = False
to arguments.
